+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Display follow request notifications.
 
+### Fixed
+
+- Fix test overflow in action element.
+
 ## [1.33.1] - 2023-08-20
 
 ### Fixed