+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Display follow request notifications.
+- Display "No posts found" label when profile timeline is empty.
 
 ### Changed