+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Changed style of "not found" messages.
 - Add FEP-8b32 integrity proof to C2S activities.
+- Don't send `params` parameter to `/api/v1/accounts/send_activity` endpoint.
 
 ### Fixed