+++ CHANGELOG.md
@@ -8,7 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-- Add "Mute author" to post dropdown menu.
+- Added "Mute author" to post dropdown menu.
+- Added OpenGraph image.
 
 ### Changed