+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Ignore response body of `/api/v1/accounts/send_activity` endpoint.
+- Improve Monero wallet login form.
 
 ### Fixed