+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Increase gap between icon and text in post menu.
 
+### Fixed
+
+- Fix vertical position of dropdown menus.
+
 ## [2.0.0] - 2023-10-02
 
 ### Removed