+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Add "Copy link to post" item to post menu.
+
 ### Changed
 
 - Increase gap between icon and text in post menu.