+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Add "Copy link to post" item to post menu.
+- Fetch subscriber profile when performing identification on Monero subscription page.
 
 ### Changed